requirements.txt
Django>=3.2
psycopg2-binary>=2.8
pandas>=1.3.0
djangorestframework>=3.12.4
openpyxl>=3.0
matplotlib>=3.4.2
mpld3>=0.5.5
plotly>=5.3.1
django-cors-headers>=3.8.0
